# api/openapi.yaml
openapi: "3.1.0"
info:
  title: OurFamRoots API
  version: "1.0.0"
  description: |
    ## OurFamRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
  parameters:
    treeId:
      name: treeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Family tree identifier
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Maximum number of items to return
    cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response

  # ──────────────────────────────────────────────────────────
  # COMMON HEADERS
  # ──────────────────────────────────────────────────────────
...
  schemas:
...
    PersonSex:
      type: string
      enum: [MALE, FEMALE, OTHER, UNKNOWN]
...
    Pagination:
      type: object
      required: [hasMore]
      properties:
        cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page (null if no more pages)
        hasMore:
          type: boolean
        totalCount:
          type: integer
          description: Total matching records (expensive; only returned when requested)

    # ── Auth / Tokens ───────────────────────────────────────
...
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /trees/{treeId}/persons:
    parameters:
      - $ref: "#/components/parameters/treeId"
    get:
      tags: [Persons]
      operationId: listPersons
      summary: List persons in a tree
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
        - name: surname
          in: query
          schema:
            type: string
          description: Filter by surname (prefix match)
        - name: sex
          in: query
          schema:
            $ref: "#/components/schemas/PersonSex"
        - name: isLiving
          in: query
          schema:
            type: boolean
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [surname, givenName, createdAt, updatedAt]
            default: surname
      responses:
        "200":
          description: Paginated list of persons
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PersonSummary"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    post:
      tags: [Persons]
      operationId: createPerson
      summary: Create a new person in a tree
      description: Requires CONTRIBUTOR, EDITOR, or ADMIN role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonRequest"
      responses:
        "201":
          description: Person created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
...
  /admin/users:
    get:
      tags: [Admin]
      operationId: listUsers
      summary: List all users in the tenant
      description: |
        Returns a paginated list of users. Requires `ADMIN` app role with a
        verified email address.

        A regular `ADMIN` stays scoped to their own namespace and never sees
        `SUPER_ADMIN` accounts. A `SUPER_ADMIN` instead sees users across
        **every** namespace by default (each item's `namespace` field is
        populated), optionally narrowed to one namespace via `namespace_id`.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
          description: >
            Search by name or email (case-insensitive). For a Super Admin,
            also matches the namespace's name/slug.
        - name: app_role
          in: query
          schema:
            type: string
            enum: [ADMIN, STANDARD, AUDITOR]
        - name: verified
          in: query
          schema: { type: boolean }
          description: Filter by email verification status
        - name: namespace_id
          in: query
          schema: { type: string, format: uuid }
          description: Super Admin only — filter to one namespace.
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at_desc, created_at_asc, name_asc, email_asc, last_login_desc]
            default: created_at_desc
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                required: [total, items, page, page_size, total_pages]
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUser"
                  page: { type: integer }
                  page_size: { type: integer }
                  total_pages: { type: integer }
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"

    post:
      tags: [Admin]
      operationId: createUser
      summary: Create a new user and send an activation email
      description: |
        Creates a user in the tenant with a random unusable password.
        Sends an activation email containing a link to `POST /auth/reset-password`
        where the user sets their own password. Email is unique across every
        namespace — an account belongs to exactly one namespace at a time.

        **Side effects:**
        - Activation email sent immediately.
        - Activity log entry created: `event_type = ADMIN_CREATE`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, given_name]
              properties:
                email:
                  type: string
                  format: email
                given_name:
                  type: string
                  minLength: 1
                  maxLength: 100
                family_name:
                  type: string
                  maxLength: 100
                  default: ""
                app_role:
                  type: string
                  enum: [ADMIN, STANDARD, AUDITOR]
                  default: STANDARD
                namespace_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: >
                    Super Admin only — create the user in this namespace
                    instead of the caller's own. Ignored for a regular ADMIN,
                    who can only create users in their own namespace.
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "409":
          description: Email already exists in tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
  /admin/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      tags: [Admin]
      operationId: updateUser
      summary: Update a user's profile, role, or verification status
      description: |
        Partial update — all body fields are optional.

        **Verification change side effects:**
        - `email_verified = true` → sends "account verified" email, logs `ADMIN_VERIFY`.
        - `email_verified = false` → sends "verification removed" email, logs `ADMIN_UNVERIFY`.

        **Activation/deactivation via this endpoint:**
        - `is_active = true` → logs `ADMIN_ACTIVATE`.
        - `is_active = false` → logs `ADMIN_DEACTIVATE`.
        - Name/role-only changes → logs `ADMIN_UPDATE`.

        **Self-protection** (when `userId` equals the caller's own ID):
        - Cannot change own `app_role` away from `ADMIN` → `400`.
        - Cannot set own `is_active = false` → `400`.
        - Cannot set own `email_verified = false` → `400`.
        - Name changes are always allowed.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                given_name:
                  type: string
                  maxLength: 100
                  nullable: true
                family_name:
                  type: string
                  maxLength: 100
                  nullable: true
                app_role:
                  type: string
                  enum: [ADMIN, STANDARD, AUDITOR]
                is_active:
                  type: boolean
                email_verified:
                  type: boolean
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: |
            Self-protection rule violated (e.g. cannot change own role or
            deactivate own account).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    delete:
      tags: [Admin]
      operationId: deactivateUser
      summary: Deactivate (soft-delete) a user
      description: |
        Sets `is_active = false` and immediately revokes all of the user's
        active refresh tokens (forcing sign-out on all devices).

        **Side effects:**
        - "Your account has been deactivated" email sent to the user.
        - Activity log entry created: `event_type = ADMIN_DEACTIVATE`.

        **Self-protection:** Cannot deactivate your own account → `400`.
      security:
        - BearerAuth: []
      responses:
        "204":
          description: User deactivated — no body
        "400":
          description: Cannot deactivate your own account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
  /admin/users/{userId}/verify:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Admin]
      operationId: adminVerifyUser
      summary: Manually verify a user's email address
      description: |
        Marks `email_verified = true` without requiring the user to click an
        email link.

        **Side effects:**
        - "Your account is verified" email sent with a sign-in link.
        - Activity log entry created: `event_type = ADMIN_VERIFY`.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
...